Adding support for firestore and auth emulators
We have been running the web client against prod, which requires credentials that we don't need in order to develop in dev.  Its also very convenient to experiment with data and blow it up when we no longer need it.  It's more difficult to do that in prod.

The only emulators we need for now are auth and firestore.  And I have deleted the configurations for all else that we currently don't use.