Moved firebase code to the backend directory
Im working on organizing the code a lil bit better so that it is easier to know what is backend and frontend code.  So I have moved all the backend code to the `backend` directory.  I have updated the instructions to run the backend and frontend code in the README.md.

But basically, you go to the `backend` directory and run `npm start` and in web-client do the same.