Added support for importing decks from archidekt
We are working on improving our ability to import decks, so I am adding support to import decks from archidekt.  The flow is pretty simple.  Go to https://www.archidekt.com/ and find a deck you want to import.  The copy the URL from the browser into magic-table new import from URL field and import the deck.  archidekt URLs look like https://www.archidekt.com/decks/6221548/rin_and_seri_cat_tribal_edh

The importing uses the oracle id to make sure we are not messing aroung card text, which may not be a reliable identifier for some older cards.

As part of these changes, I refactored some of the types to be full classes so that we aren't dealing with plain objects with types slapped on top. This is very rough right now but I will be smoothing this out in follow up changes.