Added game state and ability to add players to a game
I have added the ability to add players to a game using the player's user id.  For now that will work, but we will want to change that to username or some sort of handle that folks can use safely.

Only the owner of the game can add users.  Eventually we can add an invitation system to join games, but for now this will get us closer to being to able with other folks, even if we only see our own card.