Added fish eye zoom in different card zones
It can be difficult to see card details.  So I have added an experimental fish eye zoom that will scale up the size of cards that you hover over.  Its configurable enought that we can tweak how much things are zoomed.

I only added this to the Hand, Lands, and Battlefield zone for now.