Added better support for allowing fish eye X coordinate scroll
When there are lots of card on your hand (or other zones that list cards), fish eye didn't render much of the cards adjacent to the card in focus.  So when moving the mouse in fish eye mode would skip cards. These changes moves cards on along the X axis so that cards arent missed.