Be even smarter about excluding cards from Archidekt
Archidekt has category details about whether they should be included in the deck.  Read that data and use it when importing data instead of assuming only "Sideboard" and "Maybeboard".  Users might have created categories like "Nice to have" or "Removed" which also would be removed from the 100.

I still think we want to import "Sideboard" cards as sideboard cards in the deck.  There are cards that let you pull from the sideboard, for example.  But that is rare and we'll build that out at a later date.