Adding a Command Zone to the game board
And automatically extracting the commanders from the deck before we shuffle them, putting them in the command zone automatically.

We'll want to be smart about whether or not to include the command zone in the future.  Not all games have command zones.  But for now, it's just there and cards can be moved to/fro